feat(read-record): 重构阅读记录功能，新增阅读热力图、多维度时长统计及按日筛选，并优化数据库存储方案。

+++ app/src/main/java/io/legado/app/data/entities/ReadRecordShow.kt
@@ -1,7 +1,10 @@
 package io.legado.app.data.entities
 
+/**
+ * UI 展示用聚合数据：lastRead 为本地日期 yyyyMMdd 整数键。
+ */
 data class ReadRecordShow(
     var bookName: String,
     var readTime: Long,
-    var lastRead: Long
+    var lastRead: Int
 )
